made scandoubler configurable
In the "Audio and video settings (QNICE clock domain)" section of mega65.vhd, you can now configure if you need a scandoubler or not. Default is: OFF. You configure the scandoubler setting in the QNICE clock domain, the framework takes care of the CDC to the video_clk (which is in the current version of M2M always in phase with the main clock, so that a CDC from QNICE to main is sufficient).