added VD_ANTI_THRASHING_DELAY
Delay in ms between the last write request to a virtual drive from the core and the start of the cache flushing (i.e. writing to the SD card). Since every new write from the core invalidates the cache, and therefore leads to a completely new writing of the cache (flushing), this constant prevents thrashing. The default is 2 seconds (2000 ms). Should be reasonable for many systems, but if you have a very fast or very slow system, you might need to change this constant.