Keyboard controller supports arbitrary core clock frequencies
Up to now, the framework needed to be adjusted if the core's clock frequencies were "too different" from the MEGA65's 40 MHz. This commit solves it "for now - good enough" by adding some math to the keyboard controller. This math is being optimized by Vivado as long as the core's frequency is a constant value. For switchable frequencies (for example PAL vs. NTSC) we need a better solution. Otherwise this solution will probably generate a ton of combinatorial logic plus you will need to add some false_paths to the XDC.