QNICE: Updated to bugfixed FAT32 library
Fixed device-handle and stack corruption: If the same DEVICE handle was used multiple times in a row for f32_fopen calls that returned errors (e.g. due to file not found), then - independent from the FILE handle in use - the FAT32$DEV_BUFFERED_FDH contained a pointer to a temporary file handle that happened to have a stack address. The result: This very stack address was overwritten sometimes (which lead to stack corruptions) and the invalid "owner" of the 512-byte hardware buffer stored in FAT32$DEV_BUFFERED_FDH generated errors, for example in FAT32$RW_SIC.