Logging C_DEV_SYS_INFO more robust
We currently are only able to output pixel rates up to 65.535 MHz, which is slightly below 720p. If the result is larger, then we will output <n/a> as pixel rate. As soon as we start to work with retro cores that use such high resolutions, we need to update _LOG_DECIMAL_D. There is a @TODO in the code.