Power-LED: Updated semantics
MEGA65's power led:
Old semantics: Power led = green.
New semantics (taken from the C64 and which has proven to be helpful): By default, it is on and glows green when the MEGA65 is powered on. We switch it to blue when a long reset is detected and as long as the user keeps pressing the preset button.