Bugfix: ascal'er would sometimes freeze when switching mode
This fix is part of issue #9.

When switching modes, the HDMI clock would "flicker", i.e. have
glitches, during the transition. This caused the ascaler to freeze.

The fix is to ensure a glitch-free transition to an (arbitrary) backup
clock (of 50 MHz) while the MMCM is being re-programmed and finding its
new lock.